Added possibility to have a secondary el on the host
This change, relatively simple and backward compatible, enables syncing an external execution client from the host itself instead of relying on running within Docker. We tested it and to succesfully sync with the inner el peer, one further change is needed which is to expose the P2P address to the public interface (instead of only within the host), PR incoming for this.