Pull images in playground itself instead of docker-compose to have visibility on pull actions (#232)
Aiming to fix #213. I tried to subscribe to Docker events for image
pulling but they only happen **after** the images has been successfully
pulled (download + extract). Thus, it does not give enough visibility to
return the info to the user that the image is being pulled.

So, the solution implemented in this PR is to pull ourselves the images
before calling `docker-compose up`. This PR adds two things:
- The aforementioned support to pull not available images.
- A callback mechanism to notify about internal events happening in the
runtime execution (image being pulled, pulled, container created,
healthy, down, etc...)

This second feature is the one I want to use in a followup PR to fix
#213.