Add unit test for BuilderHub component (#238)
This PR adds a unit test for the BuilderHub component.

It introduces a new entity called `testComponent` which can deploy any
type of component in the Docker runtime and stores the execution logs
and artifacts in the `e2e-test/<time>-<test_name>` folder. I feel we can
use this framework to replace the CI e2e checks, convert them to unit
tests and expand the type of tests that we do on the recipes. That
specific work will be done in a followup PR, as of now, `testComponent`
is meant to only run one component at a time.

Closes #230