Add unit tests for the rest of recipes (#239)
Builds on top of #238.

It expands the test framework introduced in 238 for the other recipes.
The major changes are in `components_test.go`. I also replaced the e2e
tests on CI with these new tests.

There are some benefits of doing unit tests as e2e tests vs what we had
before:
- We can be more explicit and programatic about post execution test
conditions (i.e. check after the recipe is up whether transactions can
be mined).
- It can be faster because the tests can be run on parallel. Note, this
is not possible yet due to limitations on how ports are binded to the
host machine, but, right now, running all the tests takes 1.5 minutes vs
the 4 minute timeout set for each of the previous tests.
- We can run everything on the same machine.