Enable changing l1 recipe default block time (#222)
This PR introducts a `--block-time=1s` on `builder-playground cook l1`
so it's possible to product blocks faster on the launched node.

I took the libery to rename the op block time flag to be unexported and
to have the `InSeconds` suffix. I also used `time.Duration` flag type
for the `--block-time` implementation on L1 recipe, let me know if you
would like this to be applied to op block-time too, or if in the
opposite, would you prefer keeping the old way for both flag.