Add trace timer (#261)
This PR adds a `StartTimer` utility that we can use to show how much
time different internal stages of the system are taking if the
playground runs in `--log-level=trace` mode.

Example of the output:
```$ go run main.go cook l1 --output ./output --log-level trace
2025/12/17 09:08:57 Log level: trace
2025/12/17 09:08:57 deleting existing output directory ./output
2025/12/17 09:08:57 Genesis block hash: 0xf755a8f8203a2a94c3b9c028898f0c218c220bbc4a99648274726912da70e2ee
2025/12/17 09:09:00 [TRACE] artifacts.builder: 3.435216667s
2025/12/17 09:09:00 [TRACE] docker.pull-available-images: 23.35025ms
2025/12/17 09:09:18 [TRACE] docker.up: 18.081256709s
```

Note that in this case, `docker.up` contains the time for
`mev-boost-relay` to be ready and healthy which needs beacon to be
running.