Probe scripts (#260)
Playground had two systems to validate the health of a service.
- Functions in Go that would run on playground itself, used to determine
whether the whole recipe was ready or not.
- Healthcheck scripts in the docker containers managed by docker-compose
to determine dependencies in the docker-compose file.

Both systems evolved independently from each other, first it was 1, and
2 came about when some services had hard dependencies and would not
start without the others running. But, in insight they implement the
same functions, i.e. to check whether an EL node is generating blocks in
sequence and at correct block time intervals.
This PR eliminates 1 and integrates it inside 2.