Remove artifacts struct and have artifact builder take a reference to output folder (#263)
So we have this weird indirection in which we have the artifacts builder
which generates the output folder, generates the artifacts and then
returns an `Artifact` struct with that output folder to be used
elsewhere (i.e. other stages that might require creating artifacts).

To me, it feels weird that it is the artifact builder the one that owns
and creates that output folder. Instead, this PR has the artifact
builder to take a reference to the output folder.

Note how the tests and the main command have been simplified and look
much nicer with this structure.
This PR builds on top of #262